Add scripts and permissions map for packagers.
This script will read the permissions map that has been previously created
from released versions of the rootfs image. It will automatically restore
permissions that have been lost for whatever reason (including git repo
transfers).

Permissions maps can be created with the following (in a loopback-mounted
rootfs.img):

	sudo find . -exec stat -c '%a %u %g %n' {} \; >perms.map

The perms.map file should be placed in scripts/ and updated regularly when
new files are added to the image or permissions are changed.

The script must be run in the root directory of the rootfs.