Sync froyo ril blob from hyc's codebase.
This will result in a roaming indicator on FRX06 in CDMA mode.

This is a first release of an enormous RIL re-write by hyc. There will likely
be issues when using FRX06 and earlier releases. If something goes wrong which
precludes acceptable functioning with FRX06, please revert to the previous
rootfs release.